Update type hint for time variable

+++ atmosnav/agents/controllers/controller.py
@@ -11,5 +11,5 @@ class Controller(JaxTree):
     """ """
 
     @abc.abstractmethod
-    def action_to_control_input(self, time: jnp.float64, state: Array, action: Array) -> tuple[Array, Self]:
+    def action_to_control_input(self, time: jnp.float32, state: Array, action: Array) -> tuple[Array, Self]:
         """  """
